feat: register manufacturer-specific  types and messages (#4)
Registering manufacturer-specific types and messages is no longer restricted to the manufacturer's range; however, it still complies with not modifying existing types or messages.

I see some manufacturers using `Message Number` outside the range `0xFF00-0xFFFE`, e.g. `68` on Fit files generated by Bryton. The number is not specified in `Profile.xlsx`, it's probably used for internal purpose by Garmin as the creator of the Fit Protocol and the manufacturers. The ability to define it outside range will benefit the engineers using this SDK for manufacturer specific purpose.